requirements.txt
# Core dependencies
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0

# Ollama client
ollama>=0.1.0

# Reporting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Statistics
scipy>=1.11.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
